Remove TableHeader_CMR
